adding new searchFor path

+++ src/main/java/com/example/demo/exceptions/CowNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.demo.exceptions;
 
 public class CowNotFoundException extends RuntimeException {
 
-		public CowNotFoundException(Long id) {
+		public CowNotFoundException(int id) {
 			super("Could not find cow with id " + id );
 		}
 }
